+++ PinnyNotes.Core/DatabaseInitializer.cs
@@ -7,7 +7,7 @@ namespace PinnyNotes.Core;
 
 public class DatabaseInitialiser
 {
-    public const int SchemaVersion = 4;
+    public const int SchemaVersion = 5;
 
     public static async Task Initialise(string connectionString)
     {
@@ -95,12 +95,14 @@ public class DatabaseInitialiser
         Schema1To2Migration schema1To2Migration = new();
         Schema2To3Migration schema2To3Migration = new();
         Schema3To4Migration schema3To4Migration = new();
+        Schema4To5Migration schema4To5Migration = new();
 
         Dictionary<int, SchemaMigration> migrations = new()
         {
             {schema1To2Migration.TargetSchemaVersion, schema1To2Migration},
             {schema2To3Migration.TargetSchemaVersion, schema2To3Migration},
-            {schema3To4Migration.TargetSchemaVersion, schema3To4Migration}
+            {schema3To4Migration.TargetSchemaVersion, schema3To4Migration},
+            {schema4To5Migration.TargetSchemaVersion, schema4To5Migration}
         };
 
         using SqliteTransaction transaction = connection.BeginTransaction();