+++ PinnyNotes.WpfUi/Helpers/RegistryHelper.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Win32;
+using System.Reflection;
 
 namespace PinnyNotes.WpfUi.Helpers;
 
@@ -13,5 +14,24 @@ public static class RegistryHelper
         object? value = key.GetValue("AppsUseLightTheme");
 
         return value is int i && i == 0;
+    }
+
+    public static void AddToStartup()
+    {
+        using RegistryKey? key = Registry.CurrentUser.OpenSubKey(@"SOFTWARE\Microsoft\Windows\CurrentVersion\Run", writable: true);
+
+        string exePath = Environment.ProcessPath ?? Assembly.GetExecutingAssembly().Location;
+
+        key?.SetValue("PinnyNotes", $"\"{exePath}\"");
+    }
+
+    public static void RemoveFromStartup()
+    {
+        using RegistryKey? key = Registry.CurrentUser.OpenSubKey(@"SOFTWARE\Microsoft\Windows\CurrentVersion\Run", writable: true);
+
+        if (key?.GetValue("PinnyNotes") != null)
+        {
+            key.DeleteValue("PinnyNotes");
+        }
     }
 }