+++ PinnyNotes.WpfUi/Models/ApplicationSettingsModel.cs
@@ -5,6 +5,7 @@ namespace PinnyNotes.WpfUi.Models;
 public class ApplicationSettingsModel : BaseModel
 {
     // General
+    public bool StartWithWindows { get; set => SetProperty(ref field, value); }
     public StartupBehaviour StartupBehaviour { get; set => SetProperty(ref field, value); }
     public NewInstanceBehaviour NewInstanceBehaviour { get; set => SetProperty(ref field, value); }
     public bool ShowNotifyIcon { get; set => SetProperty(ref field, value); }